+++ src/shared/components/ChannelCover/ChannelCover.stories.tsx
@@ -1,17 +1,6 @@
-import { AllChannelFieldsFragment } from '@/api/queries'
 import { Meta, Story } from '@storybook/react'
 import React from 'react'
-import { Button } from '..'
 import ChannelCover, { ChannelCoverProps } from './ChannelCover'
-
-const channel: AllChannelFieldsFragment = {
-  __typename: 'Channel',
-  id: 'af8e1acc-cec4-4b3f-962f-22899f9bb617',
-  handle: 'modi sequi',
-  avatarPhotoUrl: 'https://eu-central-1.linodeobjects.com/atlas-assets/channel-avatars/2.jpg',
-  coverPhotoUrl: 'https://eu-central-1.linodeobjects.com/atlas-assets/channel-posters/2.jpg',
-  follows: 20,
-}
 
 export default {
   title: 'Shared/ChannelCover',
@@ -22,11 +11,10 @@ export default {
         required: false,
       },
     },
-    handleFollow: { action: 'followed' },
-    channel: {
-      defaultValue: {
-        ...channel,
-      },
+    handleEditCover: { action: 'edit/add cover' },
+    handleRemoveCover: { action: 'remove cover' },
+    coverPhotoUrl: {
+      defaultValue: 'https://eu-central-1.linodeobjects.com/atlas-assets/channel-posters/2.jpg',
     },
   },
 } as Meta
@@ -46,16 +34,7 @@ const Template: Story<ChannelCoverProps> = (args) => {
 
 export const Default = Template.bind({})
 
-export const WithNoImages = Template.bind({})
-WithNoImages.args = {
-  channel: {
-    ...channel,
-    avatarPhotoUrl: null,
-    coverPhotoUrl: null,
-  },
-}
-
-export const WithNoFollowButton = Template.bind({})
-WithNoFollowButton.args = {
-  handleFollow: undefined,
+export const WithNoImage = Template.bind({})
+WithNoImage.args = {
+  coverPhotoUrl: undefined,
 }
