+++ src/shared/components/ChannelCover/ChannelCover.style.ts
@@ -1,15 +1,6 @@
-import { Placeholder, Text } from '@/shared/components'
 import { breakpoints, colors, sizes, transitions, typography, zIndex } from '@/shared/theme'
 import { css } from '@emotion/react'
 import styled from '@emotion/styled'
-import { fluidRange } from 'polished'
-import Avatar from '../Avatar'
-import Button from '../Button'
-
-const SM_TITLE_HEIGHT = '44px'
-const TITLE_HEIGHT = '51px'
-const SM_SUBTITLE_HEIGHT = '24px'
-const SUBTITLE_HEIGHT = '27px'
 
 type CoverImageProps = {
   src: string
@@ -19,14 +10,6 @@ type CoverImageProps = {
 type EditableOverlayProps = {
   withImage?: boolean
 }
-
-type TitleContainerProps = {
-  editable?: boolean
-}
-
-export const Header = styled.section`
-  position: relative;
-`
 
 export const MediaWrapper = styled.div`
   margin: 0 calc(-1 * var(--global-horizontal-padding));
@@ -166,86 +149,5 @@ export const EditCoverButton = styled.button`
       border-radius: 100%;
       border: 1px solid ${colors.gray[200]};
     }
-  }
-`
-
-export const TitleSection = styled.div`
-  position: absolute;
-  margin-top: 40px;
-  top: 0;
-  width: 100%;
-  @media (min-width: ${breakpoints.small}) {
-    display: flex;
-    align-items: center;
-  }
-`
-
-export const ChannelInfo = styled.div`
-  display: flex;
-  align-items: center;
-  margin-right: ${sizes(5)};
-`
-
-export const StyledAvatar = styled(Avatar)`
-  margin: 0 ${sizes(6)} 0 0;
-  margin-bottom: ${sizes(3)};
-  z-index: 2;
-
-  @media (min-width: ${breakpoints.small}) {
-    margin: 0 ${sizes(6)} 0 0;
-  }
-`
-
-export const TitleContainer = styled.div<TitleContainerProps>`
-  z-index: 2;
-  max-width: 100%;
-  ${({ editable }) => !editable && `overflow: hidden`};
-`
-
-export const Title = styled(Text)`
-  line-height: 1;
-  padding: ${sizes(1)} ${sizes(2)} ${sizes(2)};
-  ${fluidRange({ prop: 'fontSize', fromSize: '32px', toSize: '40px' })};
-  background-color: ${colors.gray[800]};
-  text-overflow: ellipsis;
-  max-height: 250px;
-  overflow: hidden;
-  @media (min-width: ${breakpoints.small}) {
-    word-break: break-all;
-  }
-`
-
-export const SubTitle = styled(Text)`
-  padding: ${sizes(1)} ${sizes(2)};
-  ${fluidRange({ prop: 'fontSize', fromSize: '14px', toSize: '18px' })};
-  margin-top: ${sizes(2)};
-  color: ${colors.white};
-  background-color: ${colors.gray[800]};
-  display: inline-block;
-`
-
-export const TitlePlaceholder = styled(Placeholder)`
-  width: 300px;
-  height: ${SM_TITLE_HEIGHT};
-  @media screen and (min-width: ${breakpoints.medium}) {
-    height: ${TITLE_HEIGHT};
-  }
-`
-
-export const SubTitlePlaceholder = styled(Placeholder)`
-  width: 140px;
-  margin-top: ${sizes(2)};
-  height: ${SM_SUBTITLE_HEIGHT};
-  @media screen and (min-width: ${breakpoints.medium}) {
-    height: ${SUBTITLE_HEIGHT};
-  }
-`
-export const StyledButton = styled(Button)`
-  position: relative;
-  z-index: 3;
-  margin-top: ${sizes(2)};
-  @media screen and (min-width: ${breakpoints.small}) {
-    margin-top: 0;
-    margin-left: auto;
   }
 `
